+++ HISTORY.md
@@ -2,6 +2,7 @@
 * [Add] Добавлен сниппет DLBuildMenu
 * [Add] Добавлен сниппет DLPrevNext
 * [Add] Добавлен хелпер класс DLHelper
-* [Add] Библиотека MODxAPI перенесена в репозиторий DocLister
+* [Refactor] Библиотека MODxAPI перенесена в репозиторий DocLister
+* [Refactor] Модуль RelativeTVList перемещен в репозиторий DocLister
 * [Refactor] Добавлена поддержка микроформата Breadcrumb schema.org
 * [Refactor] Пересмотрена структура каталогов