+++ HISTORY.md
@@ -2,9 +2,11 @@
 * [Add] Добавлен сниппет DLBuildMenu
 * [Add] Добавлен сниппет DLPrevNext
 * [Add] Добавлен хелпер класс DLHelper
+* [Add] Добавлена возможность подмены шаблона для обрабатываемого документа через переменную DocLister::renderTPL
 * [Refactor] Библиотека MODxAPI перенесена в репозиторий DocLister
 * [Refactor] Модуль RelativeTVList перемещен в репозиторий DocLister
 * [Refactor] Модуль RedirectMap2 перемещен в репозиторий DocLister
 * [Refactor] Плагин getPageID перемещен добавлен DocLister
 * [Refactor] Добавлена поддержка микроформата Breadcrumb schema.org
 * [Refactor] Пересмотрена структура каталогов
+* [Refactor] Перенос сниппета Summary в репозиторий DocLister